+++ src/page-objects/AdministrationPages.ts
@@ -45,6 +45,7 @@ import { Media } from './administration/Media';
 import { YourProfile } from './administration/YourProfile';
 import { ThemesListing } from './administration/ThemesListing';
 import { ThemesDetail } from './administration/ThemesDetail';
+import { MediaListing } from './administration/MediaListing';
 
 export interface AdministrationPageTypes {
     AdminProductDetail: ProductDetail;
@@ -91,6 +92,7 @@ export interface AdministrationPageTypes {
     AdminYourProfile: YourProfile;
     AdminThemesListing: ThemesListing;
     AdminThemesDetail: ThemesDetail;
+    AdminMediaListing: MediaListing;
 }
 
 export const AdminPageObjects = {
@@ -138,6 +140,7 @@ export const AdminPageObjects = {
     YourProfile,
     ThemesListing,
     ThemesDetail,
+    MediaListing,
 }
 
 export const test = base.extend<FixtureTypes>({
@@ -312,7 +315,12 @@ export const test = base.extend<FixtureTypes>({
     AdminThemesListing: async ({ AdminPage }, use) => {
         await use(new ThemesListing(AdminPage));
     },
+
     AdminThemesDetail: async ({ AdminPage }, use) => {
         await use(new ThemesDetail(AdminPage));
+    },
+
+    AdminMediaListing: async ({ AdminPage }, use) => {
+        await use(new MediaListing(AdminPage));
     },
 });