+++ tests/PageObjects/AdministrationGeneral.spec.ts
@@ -1,6 +1,4 @@
 import { test } from '../../src';
-import { satisfies } from 'compare-versions';
-
 
 test('Administration page objects - General.', async ({
     InstanceMeta,
@@ -75,16 +73,4 @@ test('Administration page objects - General.', async ({
     await AdminDashboard.adminMenuUserActions.click();
     await ShopAdmin.expects(AdminDashboard.adminMenuLogoutButton).toBeVisible();
 
-    // eslint-disable-next-line playwright/no-conditional-in-test
-    if (satisfies(InstanceMeta.version, '>=6.7.1')) {
-        await ShopAdmin.goesTo(AdminShopwareServices.url());
-        await ShopAdmin.expects(AdminShopwareServices.header).toBeVisible();
-        await ShopAdmin.expects(AdminShopwareServices.deactivateServicesButton).toBeVisible();
-        await AdminShopwareServices.deactivateServicesButton.click();
-        await ShopAdmin.expects(AdminShopwareServices.deactivateServicesModal).toBeVisible();
-        await ShopAdmin.expects(AdminShopwareServices.deactivateServicesConfirmButton).toBeVisible();
-        await AdminShopwareServices.deactivateServicesConfirmButton.click();
-        await ShopAdmin.expects(AdminShopwareServices.activateServicesButton).toBeVisible();
-        await AdminShopwareServices.activateServicesButton.click();
-    }
 });