+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [11.29.1](https://github.com/shopware/acceptance-test-suite/compare/v11.29.0...v11.29.1) (2025-09-10)
+
+
+### Bug Fixes
+
+* disable services locator tests in SaaS ([#454](https://github.com/shopware/acceptance-test-suite/issues/454)) ([1c80756](https://github.com/shopware/acceptance-test-suite/commit/1c8075690976a83a96085387b190a4052c40d8e9))
+* empty state selector ([#458](https://github.com/shopware/acceptance-test-suite/issues/458)) ([853981f](https://github.com/shopware/acceptance-test-suite/commit/853981f5b3d47611c2452a0313d7f21f4c41bcf7))
+
+## [11.29.0](https://github.com/shopware/acceptance-test-suite/compare/v11.28.0...v11.29.0) (2025-09-08)
+
+
+### Features
+
+* replace elements with individual replacements ([#466](https://github.com/shopware/acceptance-test-suite/issues/466)) ([0543da6](https://github.com/shopware/acceptance-test-suite/commit/0543da6ced7f48e5a34fda671914c9be10630402))
+
 ## [11.28.0](https://github.com/shopware/acceptance-test-suite/compare/v11.27.0...v11.28.0) (2025-09-05)